Implement verify workflow stage
Linting and folder structure checking.

The rules for folder structure checking are:
In src there must be only main and test folders

main

No .h .cpp or .c files in this folder (not including subfolders)

irioCore

All .cpp and .h files must start with irio and be followed by another word starting with an upper case (irio[A-Z][a-zA-Z0-9]+.(cpp|h)$)

irioCoreCpp

includes

Only .h in this folder and in all subfolders

profiles

Must start with profile and follow with another word starting with a capital letter (profile(s|[A-Z])[a-zA-Z0-9]*.h$) with the exception of "profiles.h".

terminals

Must begin with terminals and be followed by another word beginning in uppercase (terminals([A-Z][a-zA-Z0-9]*)?.h$)

impl

Must begin with terminals and follow with another word that begins with an uppercase letter and end with Impl (terminals[A-Z][a-zA-Z0-9]+Impl.h$).

All Impl must have a corresponding non-Impl in the folder above

names

Must start with namesTerminals and continue with another word that starts with a capital letter (namesTerminals[A-Z][a-zA-Z0-9]+.h$)

profiles

Must start with profile and follow with another word that begins in uppercase (profile[A-Z][a-zA-Z0-9]*.cpp$)

terminals

Must begin with terminals and follow with another word beginning in upper case (terminals[A-Z][a-zA-Z0-9]*.cpp$)

impl

Must begin with terminals and follow with another word that begins with an uppercase letter and end with Impl (terminals[A-Z][a-zA-Z0-9]+Impl.cpp$).

All Impl must have a corresponding non-Impl in the folder above

NiFpga_CD

You must only have two files (only counting those that are .c or .h)

NiFpga.c

NiFpga.h