Modify Makefiles to improve parallel compilation
The recipe in charge of updating the library version makes all the prerequsiites to run serially. Making slower to compile.

The solution has been adding this recipe as prerequisite of the cpp and c recipes. Which make them wait until the update version recipe has ended and then it runs in parallel the compilation.