Modify Makefile to avoid recompiling when nothing has change
Previously, the update version recipe, which is a PHONY recipe, was as a prerequisite for .o and .cpp files, which made it so all files must be recompiled. Now, it has been removed from prerequisite and that recipe has as a prerequisite the file that needs to be updated, which allows to compile much faster.